requirements.txt
livekit-agents~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-groq~=1.0
livekit-plugins-google~=1.0
livekit-plugins-cartesia~=1.0
livekit-plugins-deepgram~=1.0
livekit-plugins-silero~=1.0
# optional, only when using LiveKit's turn detection model
livekit-plugins-turn-detector~=1.0
# optional, only if background voice & noise cancellation is needed
livekit-plugins-noise-cancellation>=0.2.0,<1.0.0
python-dotenv~=1.0
flask~=2.3.0
requests~=2.31.0
loguru
webdriver-manager
selenium
pandas
gunicorn